Improve robustness of setup_external against errors and reruns

+++ setup_external.sh
@@ -1,12 +1,16 @@
+set -e
+
 # ===== PREP =====
 mkdir -p external
 pushd external
 setup_root=$(pwd)
 
 # ===== VIENNA (1.x) =====
-wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/1_8_x/ViennaRNA-1.8.5.tar.gz
-tar -xvf ViennaRNA-1.8.5.tar.gz
-rm ViennaRNA-1.8.5.tar.gz
+if [ ! -d ViennaRNA-1.8.5 ]; then
+    wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/1_8_x/ViennaRNA-1.8.5.tar.gz
+    tar -xvf ViennaRNA-1.8.5.tar.gz
+    rm ViennaRNA-1.8.5.tar.gz
+fi
 pushd ViennaRNA-1.8.5
 
 ./configure CFLAGS="-std=gnu89 -fcommon -g -O2" --without-perl --without-forester --without-kinfold --prefix=$setup_root/ViennaRNA-1.8.5/build
@@ -17,9 +21,11 @@ make install
 popd
 
 # ===== VIENNA (2.1.9) =====
-wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/2_1_x/ViennaRNA-2.1.9.tar.gz
-tar -xvf ViennaRNA-2.1.9.tar.gz
-rm ViennaRNA-2.1.9.tar.gz
+if [ ! -d ViennaRNA-2.1.9 ]; then
+    wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/2_1_x/ViennaRNA-2.1.9.tar.gz
+    tar -xvf ViennaRNA-2.1.9.tar.gz
+    rm ViennaRNA-2.1.9.tar.gz
+fi
 pushd ViennaRNA-2.1.9
 
 ./configure --prefix=$setup_root/ViennaRNA-2.1.9/build --without-perl --without-forester --without-kinfold
@@ -29,9 +35,11 @@ make install
 popd
 
 # ===== VIENNA (2.6.4/Current Latest) =====
-wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/2_6_x/ViennaRNA-2.6.4.tar.gz
-tar -xvf ViennaRNA-2.6.4.tar.gz
-rm ViennaRNA-2.6.4.tar.gz
+if [ ! -d ViennaRNA-2.6.4 ]; then
+    wget https://www.tbi.univie.ac.at/RNA/download/sourcecode/2_6_x/ViennaRNA-2.6.4.tar.gz
+    tar -xvf ViennaRNA-2.6.4.tar.gz
+    rm ViennaRNA-2.6.4.tar.gz
+fi
 pushd ViennaRNA-2.6.4
 
 ./configure --prefix=$setup_root/ViennaRNA-2.6.4/build --without-perl --without-python --without-forester --without-kinfold --without-rnalocmin --without-rnaxplorer
@@ -42,8 +50,13 @@ make install
 popd
 
 # ===== NEMO =====
-git clone https://github.com/eternagame/nemo
-pushd nemo
+if [ ! -d nemo ]; then
+    git clone https://github.com/eternagame/nemo
+    pushd nemo
+else
+    pushd nemo
+    git fetch
+fi
 git checkout 154ada38c33efa4ccb1502f368a790d1b8124381
 
 gcc -g -O2 -fopenmp -I../ViennaRNA-2.1.9/H -MD -MP -c -o nemo.o nemo.cpp
@@ -52,8 +65,17 @@ gcc -g -O2 -fopenmp -L../ViennaRNA-2.1.9/build/lib -o nemo nemo.o -lstdc++ -lRNA
 popd
 
 # ===== ETERNABRAIN =====
-git clone https://github.com/eternagame/eternabrain
-pushd eternabrain
+if [ ! -d eternabrain ]; then
+    git clone https://github.com/eternagame/eternabrain
+    pushd eternabrain
+else
+    pushd eternabrain
+    git fetch
+fi
+# In order to make sure our patch applies cleanly, throw out any prior modifications
+# We use git stash instead of git clean just in case there were intentional local modifications
+# someone wants to retrieve (the overhead incurred here should not be substantial)
+git stash -u
 git checkout 3b152333b579ef49ca843c497b3f11045e67b319
 git apply ../eternabrain.patch
 
@@ -64,15 +86,30 @@ conda create -qy -p ../eternabrain-env python=2.7
 popd
 
 # ===== LEARNA =====
-git clone https://github.com/automl/learna
-pushd learna
+if [ ! -d learna ]; then
+    git clone https://github.com/automl/learna
+    pushd learna
+else
+    pushd learna
+    git fetch
+fi
+# In order to make sure our patch applies cleanly, throw out any prior modifications
+# We use git stash instead of git clean just in case there were intentional local modifications
+# someone wants to retrieve (the overhead incurred here should not be substantial)
+git stash -u
 git checkout 34b3cf9eeaab848a7fa8b1481b4abe0c416fe09e
 git apply ../learna.patch
 
 conda env create -qy -p ../learna-env --file environment.yml
 
 # ===== SENTRNA =====
-git clone https://github.com/jadeshi/SentRNA
+if [ ! -d SentRNA ]; then
+    git clone git clone https://github.com/jadeshi/SentRNA
+    pushd SentRNA
+else
+    pushd SentRNA
+    git fetch
+fi
 pushd SentRNA
 git checkout 6855a2d0734a962ffa7bf5a5833f52695644e18c
